html/testdata/go/doctype_named_entity.dat

#data
<!DOCTYPE &gt; PUBLIC "&gt;" "&gt;">
#errors
#document
| <!DOCTYPE > ">" ">">
| <html>
|   <head>
|   <body>
